# chore: improve coverage for the case that the cipher that is not long enough