<!--
Thank you for your pull request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests and possibly benchmarks.

Contributors guide: https://github.com/fastify/fastify/blob/master/CONTRIBUTING.md

By making a contribution to this project, I certify that:

* (a) The contribution was created in whole or in part by me and I
  have the right to submit it under the open source license
  indicated in the file; or

* (b) The contribution is based upon previous work that, to the best
  of my knowledge, is covered under an appropriate open source
  license and I have the right under that license to submit that
  work with modifications, whether created in whole or in part
  by me, under the same open source license (unless I am
  permitted to submit under a different license), as indicated
  in the file; or

* (c) The contribution was provided directly to me by some other
  person who certified (a), (b) or (c) and I have not modified
  it.

* (d) I understand and agree that this project and the contribution
  are public and that a record of the contribution (including all
  personal information I submit with it, including my sign-off) is
  maintained indefinitely and may be redistributed consistent with
  this project or the open source license(s) involved.
-->

#### Checklist

- [x] run `npm run test` and `npm run benchmark`
- [x] tests and/or benchmarks are included
- [ ] documentation is changed or added
- [x] commit message and code follows the [Developer's Certification of Origin](https://github.com/fastify/.github/blob/master/CONTRIBUTING.md#developers-certificate-of-origin-11)
      and the [Code of conduct](https://github.com/fastify/.github/blob/master/CODE_OF_CONDUCT.md)

before:
```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 41.229ms
 PASS  test/cookie-name.js 6 OK 168.487ms
 PASS  test/key-base64.js 5 OK 168.367ms
 PASS  test/delete.js 14 OK 166.961ms
 PASS  test/access.js 20 OK 205.971ms
 PASS  test/key-rotation.js 11 OK 183.691ms
 PASS  test/dynamic-cookie-options.js 13 OK 174.026ms
 PASS  test/key.js 12 OK 198.419ms
 PASS  test/malformed-cookie.js 5 OK 141.524ms
 PASS  test/nonce-bad-length.test.js 5 OK 112.986ms
 PASS  test/salt.js 5 OK 116.898ms
 PASS  test/path.js 4 OK 117.56ms
 PASS  test/schema.js 6 OK 143.425ms
 PASS  test/secret.js 21 OK 2s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  136 passed, of 136
Time:   4s
ERROR: Coverage for lines (95.96%) does not meet global threshold (100%)
ERROR: Coverage for branches (95.31%) does not meet global threshold (100%)
ERROR: Coverage for statements (96%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|---------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|---------------------
All files |      96 |    95.31 |     100 |   95.96 |
 index.js |      96 |    95.31 |     100 |   95.96 | 118-119,141-142,262
----------|---------|----------|---------|---------|---------------------
```

after:
```
$ npm test

> @fastify/secure-session@5.1.0 test
> standard && tap test/*.js && tsd

 PASS  test/decorators.js 9 OK 362.561ms
 PASS  test/delete.js 14 OK 163.059ms
 PASS  test/cookie-name.js 6 OK 160.86ms
 PASS  test/key-base64.js 5 OK 155.412ms
 PASS  test/access.js 20 OK 185.963ms
 PASS  test/key-rotation.js 11 OK 197.702ms
 PASS  test/key.js 12 OK 187.271ms
 PASS  test/dynamic-cookie-options.js 13 OK 174.655ms
 PASS  test/malformed-cookie.js 5 OK 124.065ms
 PASS  test/path.js 4 OK 128.767ms
 PASS  test/salt.js 5 OK 130.885ms
 PASS  test/nonce-bad-length.test.js 5 OK 115.503ms
 PASS  test/schema.js 6 OK 132.452ms
 PASS  test/secret.js 21 OK 2s



  🌈 SUMMARY RESULTS 🌈


Suites:   14 passed, 14 of 14 completed
Asserts:  136 passed, of 136
Time:   4s
ERROR: Coverage for lines (97.58%) does not meet global threshold (100%)
ERROR: Coverage for branches (96.87%) does not meet global threshold (100%)
ERROR: Coverage for statements (97.6%) does not meet global threshold (100%)
----------|---------|----------|---------|---------|-------------------
File      | % Stmts | % Branch | % Funcs | % Lines | Uncovered Line #s
----------|---------|----------|---------|---------|-------------------
All files |    97.6 |    96.87 |     100 |   97.58 |
 index.js |    97.6 |    96.87 |     100 |   97.58 | 141-142,262
----------|---------|----------|---------|---------|-------------------
```